Create FlowerGrid.cpp
While strolling in a garden, you come across a pattern of flowers. It goes as follows:
The shape of the pattern is a square.
The pattern consists of variety of flowers. Let’s represent each type of flower by an alphabet.
If a single flower type is used, then we use ‘A’to represent the flower and the pattern is a 2 * 2 grid with all ‘A’s:
AA
AA
With every additional flower type used, each of the dimensions increases by 2.
The next alphabet letter is chosen to represent the type of flower. This variety goes around the boundary of the preceding pattern. So, if two kinds of flowers are used then the pattern is a 4 x 4 grid which looks like below:
BBBB
BAAB
BAAB
BBBB

Sample Case O
Sample Input For Custom Testing
1
Sample Output
AA
AA
Explanation
Since the number of flower types to be used is 1, the pattern consists of all As in a 2×2 grid and hence output string is “AA\nAA”

Sample Case 1
Sample Input For Custom Testing
3
Sample Output
CCCCCC
CBBBBC
CBAABC
CBAABC
CBBBBC
CCCCCC